+++ doc/release-notes.rst
@@ -232,9 +232,14 @@ Major Changes from Kraken
     - ``ceph osd reweightn`` will specify the `reweight` values for
       multiple OSDs in a single command.  This is equivalent to a series of
       ``ceph osd reweight`` commands.
-    - ``ceph osd crush class {create,rm,ls,rename}`` manage the new
+    - ``ceph osd crush class {rm,ls,ls-osd}`` manage the new
       CRUSH *device class* feature.  ``ceph crush set-device-class
       <class> <osd> [<osd>...]`` will set the class for particular devices.
+      Note that if you specify a non-existent class, it will be created
+      automatically. ``ceph crush rm-device-class <osd> [<osd>...]``
+      will instead remove the class for particular devices.
+      And if a class contains no more devices, it will be automatically
+      destoryed.
     - ``ceph osd crush rule create-replicated`` replaces the old
       ``ceph osd crush rule create-simple`` command to create a CRUSH
       rule for a replicated pool.  Notably it takes a `class` argument